chore: move etherwarp stuff in Etherwarp
affected modules: EtherwarpOverlay, EtherwarpSound and LCEtherwarp
